CityScrapper has the responsability to get the information from the city card
